requirements.txt
flask
paho-mqtt
authlib
python-dotenv
six
